IAM policy: Admin - tighten/make consistent
Enforcing a consistent separation between the right to schedule backups and the
right to tag backups for deletion, because there is nothing I can do to prevent
inadvertent use of the *Administer policies for automations (which could be
hacked to take advantage of the prior exception), or to untrustworthy
individuals (who, short of deleting all backups, could simply tag them and let
some future automation delete them all). Too risky!